requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
termcolor==1.1.0
typing_extensions==4.1.1
yarl==1.7.2
